save T.14

+++ 9-binary_tree_height.c
@@ -6,8 +6,8 @@
  */
 size_t binary_tree_height(const binary_tree_t *tree)
 {
-	int rh;
-	int lh;
+	size_t rh;
+	size_t lh;
 
 	if (!tree)
 	return (0);
